+++ examples/fastapi_full/__init__.py
@@ -1,5 +1,5 @@
 """fastapi_full fastapi app"""
 
-__version__ = "2.2.2"
+__version__ = "2.2.3"
 
 __all__ = ["__version__"]